+++ packages/ui-patterns/index.tsx
@@ -17,6 +17,7 @@ export * from './Dialogs/ConfirmDialog'
 export * from './Dialogs/ConfirmationModal'
 export * from './AssistantChat/AssistantChatForm'
 export * from './AssistantChat/AssistantCommandsPopover'
+export * from './Sheet'
 export * from './SqlToRest'
 export * from './PromoToast'
 export * from './admonition'
